# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Harmoniq project",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of harmoniq project"
  },
  "servers": [
    {
      "url": "https://truestack.onrender.com"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Register a user",
        "operationId": "registerUser",
        "description": "User registration with required name, email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John",
                    "description": "User's name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john12345@gmail.com",
                    "description": "User's email"
                  },
                  "password": {
                    "type": "string",
                    "example": "12345john",
                    "description": "User's password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/201"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Email is already in use"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "operationId": "loginUser",
        "description": "Login with email and password, receive an access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john12345@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "12345john"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Successfully logged in user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in an user!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "NuB85QNMA+UUUZWUdSPZ+mRyeVIaJfAkONkajoXn"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string",
                              "example": "John",
                              "description": "User's name"
                            },
                            "email": {
                              "type": "string",
                              "format": "email",
                              "example": "john12345@gmail.com",
                              "description": "User's email"
                            },
                            "_id": {
                              "type": "string",
                              "example": "65ca67e7ae7f10c88b598384",
                              "description": "User ID"
                            },
                            "avatar": {
                              "type": "string",
                              "format": "uri",
                              "example": "https://res.cloudinary.com/dgce49ndc/image/upload/v1752447915/ar3ps6z0sbxexlj7spgg",
                              "description": "URL to user's avatar image"
                            },
                            "savedArticles": {
                              "type": "array",
                              "description": "List of user`s saved articles"
                            },
                            "articlesAmount": {
                              "type": "integer",
                              "description": "number of articles added by the user",
                              "example": 4
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404UserNotFound"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout user",
        "operationId": "logoutUser",
        "description": "Logs out a user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully logged out, no content"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
      "201": {
        "description": "Successful registration of a new user",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 201
                },
                "message": {
                  "type": "string",
                  "example": "Successfully registered a user!"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "_id",
                    "name",
                    "email",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "65ca67e7ae7f10c88b598384",
                      "description": "User ID"
                    },
                    "name": {
                      "type": "string",
                      "example": "John",
                      "description": "User's name"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "john12345@gmail.com",
                      "description": "User's email"
                    },
                    "avatarUrl": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://res.cloudinary.com/dgce49ndc/image/upload/v1752447915/ar3ps6z0sbxexlj7spgg",
                      "description": "URL to user's avatar image"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-07-26T12:18:02.095Z",
                      "description": "User creation date"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-07-26T12:18:02.095Z",
                      "description": "User update date"
                    },
                    "accessToken": {
                      "type": "string",
                      "description": "Access token",
                      "example": "NuB85QNMA+UUUZWUdSPZ+mRyeVIaJfAkONkajoXn"
                    },
                    "savedArticles": {
                      "type": "array",
                      "description": "List of user`s saved articles"
                    },
                    "articlesAmount": {
                      "type": "integer",
                      "description": "Number of articles added by the user",
                      "example": 4
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "404UserNotFound": {
        "description": "NotFound",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "error"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                },
                "error": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}